rename componetent again :)
